additional testing added
