--- npm-package/jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  testEnvironment: 'node'
-};
